Add files via upload
The program file named PDA.nb, written in Mathematica, contains two functions, Optimize[] and Discriminate[]. The function, Optimize[] creates a dichotomous predictive function to judge if a surgery of a patient would be difficult to perform or not, using normal-based linear discriminant rule. The other function, Discriminate[], judges the difficulty of a surgery of a new patient. The practical example of these functions are written in a file, Run Sample B=20.nb. The detailed explanation about these functions are written in Read Me First.nb.